fix: stop config/IPC holes from panicking the session
Growing tags_length on reload left Pertag short, so close/view indexed off the end. Cap n_master and signed geometry on the live set_config path the same way file load and persist already do, implement the advertised focus_tab IPC, and keep Wayland motion in i32 so wide desktops do not wrap.

+++ src/jwm/input_handler.rs
@@ -3376,8 +3376,8 @@ impl Jwm {
         &mut self,
         backend: &mut dyn Backend,
         _window: Option<WindowId>,
-        root_x: i16,
-        root_y: i16,
+        root_x: i32,
+        root_y: i32,
         _time: u32,
     ) -> Result<(), Box<dyn std::error::Error>> {
         // A left press on a tab cell turns into a reorder drag once the
